Lots of stability fixes and a few more minor improvements. Timeshifting, recording and scanning now works again. Next steps are to: - change FTA to a tri-state value - change LCN to a string - fix ATSC LCN handling (remove maj/min channel, fix scan unknown channel name function) - wire in the new encryption analyser - modify scanning code to support adding channels from SDT other - add scanning settings to avoid waiting for BAT and/or NIT to reduce scan times - add min/max scan time settings - use MP modulation set - downloadable tuning details for DVB-T, C and IP - add cell ID - add stream monitoring mechanisms - add support for remembering deleted channels

+++ TvEngine3/TVLibrary/TvLibrary.Interfaces/DvbConditionalAccess.cs
@@ -2084,9 +2084,9 @@ namespace TvLibrary.Interfaces
       {
         PmtElementaryStream es = new PmtElementaryStream();
         es.StreamType = (StreamType)data[offset++];
-        if (Enum.IsDefined(typeof(LogicalStreamType), es.StreamType))
+        if (Enum.IsDefined(typeof(LogicalStreamType), (int)es.StreamType))
         {
-          es.LogicalStreamType = (LogicalStreamType)es.StreamType;
+          es.LogicalStreamType = (LogicalStreamType)(int)es.StreamType;
         }
         else
         {
